requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
neo4j==5.13.0
python-dotenv==0.19.0
pydantic==1.8.2
email-validator==1.1.3
pytest==7.4.0
httpx==0.24.1
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.5